Refactor
pkg/adapterTesting -> pkg/adapter/testing
