+++ adapter/genericListChecker/adapter_test.go
@@ -18,7 +18,7 @@ import (
 	"testing"
 
 	"istio.io/mixer/adapter/genericListChecker/config"
-	"istio.io/mixer/pkg/adapterTesting"
+	at "istio.io/mixer/pkg/adapter/testing"
 )
 
 type testCase struct {
@@ -83,5 +83,5 @@ func TestAll(t *testing.T) {
 }
 
 func TestInvariants(t *testing.T) {
-	adapterTesting.TestAdapterInvariants(newAdapter(), Register, t)
+	at.TestAdapterInvariants(newAdapter(), Register, t)
 }
