+++ adapter/stdioLogger/BUILD
@@ -21,7 +21,7 @@ go_test(
     srcs = ["stdioLogger_test.go"],
     library = ":go_default_library",
     deps = [
-        "//pkg/adapterTesting:go_default_library",
+        "//pkg/adapter/testing:go_default_library",
         "@com_github_google_go_genproto//googleapis/rpc/status:go_default_library",
     ],
 )
