+++ pkg/adapter/testing/BUILD
